pyproject.toml
[project]
name = "meteora-lp-agent"
version = "0.1.0"
description = "Autonomous Meteora DLMM LP strategy optimizer"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "httpx>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
